parse data correctly added 1-4

+++ src/utils/util.py
@@ -288,8 +288,8 @@ def parse_data(sample, rate=0.2, is_test=False, length=100, include_features=Non
         obs_data = np.nan_to_num(evals, copy=True)
         obs_data = obs_data.reshape(shp)
         # obs_intact = np.nan_to_num(obs_intact, copy=True)
-    mask = mask.float()
-    obs_mask = obs_mask.float()
+    mask = mask.astype(float)
+    obs_mask = obs_mask.astype(float)
     target_mask = obs_mask - mask
     return obs_data, obs_mask, mask, target_mask
